Creating 16x16 grid using javascript DOM
With the javascript file, helper methods were creted to generate a
square grid by row and columns. The dimmension will depend based on
the value of the constant variable called 'DIMMENSION'. Classes from
the CSS file were also manually added to elements in the javascript
file. The entire grid is then added to the body of the html.