Added layout for non-functional generate button
A top div with a button inside was added directly in the index file.
With CSS, it was styled and the margin changed to have a clean layout.
The button does not function at the moment, but it will be used to
generate and regenerate a grid by putting input into the grid.