Added functionality to the generate button
The grid's creation code was moved to an event upon clicking on the
generate button. Instead of being a predefined dimmension, the grid's
dimmensions will be determined by the user input upon clicking.

At the moment, the button does not validate the user's input.

At the moment, the button does not delete the previous grid that was
on the grid (if there was one present).

At the moment, the grid's size grows if a large dimmensions is given
instead of shrinking the squares.